add inventory + new element for projectActions
