+++ src/styles/ProjectActions.css
@@ -1,0 +1,22 @@
+@import 'variables.css';
+
+.ProjectActions {
+  display: flex;
+  justify-content: space-around;
+  flex-grow: 1;
+  align-items: flex-end;
+  padding: 3em;
+
+  &-action {
+    cursor: pointer;
+  }
+
+  /* placeholder pending images */
+  .dummyButton {
+    height: 30px;
+    width: 30px;
+    margin: 0 auto;
+    background-color: var(--colors-grey);
+    display: block;
+  }
+}
